Separate switch and shutter decoder implementations
